Moving the ImageUriProvider to core, adding fileAuthority as dependency
