+++ dribbble/src/main/java/io/plaidapp/dribbble/dagger/DribbbleModule.kt
@@ -24,6 +24,7 @@ import io.plaidapp.core.dagger.dribbble.DribbbleDataModule
 import io.plaidapp.core.data.CoroutinesContextProvider
 import io.plaidapp.core.dribbble.data.ShotsRepository
 import io.plaidapp.core.ui.widget.ElasticDragDismissFrameLayout
+import io.plaidapp.dribbble.BuildConfig
 import io.plaidapp.dribbble.domain.GetShareShotInfoUseCase
 import io.plaidapp.dribbble.ui.shot.ShotActivity
 import io.plaidapp.dribbble.ui.shot.ShotViewModel
@@ -65,4 +66,7 @@ class DribbbleModule(private val activity: ShotActivity, private val shotId: Lon
             coroutinesContextProvider
         )
     }
+
+    @Provides
+    fun fileAuthority() = BuildConfig.FILES_AUTHORITY
 }
